# managers/src/main/go/com.github/docs/swagger.yaml
basePath: /
definitions:
...
  dto.ManagerDto:
    properties:
      firstName:
        example: Michal
        type: string
      lastName:
        example: Mackonagen
        type: string
    type: object
...
host: localhost:8887
info:
  contact: {}
  description: This is a sample service for managers
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Managers API
  version: "1.0"
paths:
...
  /v1/fetch/{orderId}:
    get:
      consumes:
      - application/json
      description: Get manager by order id
      parameters:
      - description: Order ID
        in: path
        name: status
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.ManagerDto'
        "400": {}
        "403": {}
      summary: Get manager by order id
      tags:
      - manager
swagger: "2.0"